Refinement of read write optimization. At the start of a constructor, 'this' is considered to be read and written.
Also, added an opt target to ant for using -O when compiling the runtime. That is, use 'ant opt fabric-runtime' to compile with optimization.

Note that this optimization can't be used with the current multi-threaded client (this client needs all reads and writes to be checked when there are multiple threads).