Fixed the bootstrapping problem for core principals.
Cores now self-initialize; GenMap is no longer needed.  On initialization, the naming map is created at onum 0 and the core principal is created at onum 1.

If there are multiple cores hosted on a node, the principal on the first core mentioned in the node configuration will be used as the client principal.

Refactored ObjectStore and its subclasses a bit.

Removed CoreClient via some refactoring -- it was mostly a copy of Client anyway.

Removed a deprecated constructor for Object.$Impl.  It was no longer being used.