Checkpoint: refactored stuff in preparation for encryption.
------------------------------------

Clients now establish up to two separate connections to each core -- one for use as a dissemination node, the other for use as a regular client node.

Turned dissemination back on.

Added a DisseminationReadMessage for use by dissemination nodes.  This is different from a ReadMessage because the access control checks are different, and the data returned is encrypted and signed.  (No encryption or signing is done yet.)

Disentangled Globs and ObjectGroups.  ObjectGroups are unencrypted groups of objects (these used to be called Globs).  Globs are now encrypted and signed groups objects.  (No encryption or signing is done yet.)

Removed StoreException.  It was confusing and served the exact same purpose as AccessException.