JifContext requires a jl type system to find "jif.lang.Principal" and classes that represent principals. In fabric, we should still give it a jl type system, rather than a fabil type system.
Now fabric signatures all build, without any special trick in build.xml (no need to do it in two passes).

Added FabILTypeBuilder, which checks for duplicate classes. It is probably not necessary with the above fix and the fix in polyglot. (Possibly neither is FabricTypeBuilder). But it doesn't hurt to leave it there...