Enabled remote calls.
Clients now specify in their properties file the port on which they listen for remote-call requests.

Clients publish their remote-call port numbers in DNS.  Hacked the NameService to simulate this.

Added a shutdown mechanism to RemoteCallManager.

Changed dissemination to use higher ports.  This avoids conflicts with the remote call manager.