Bugfixes and tweaks.  Minor build refactoring. Now compiles all of FabIL runtime.
I moved the array classes into a separate directory to make it more clear that
they should not be part of the sourcepath when compiling the runtime.
Previously, they got picked up in the sourcepath and we had a weird
resolver hack to ensure we got the right type for code generation.