Added ClassHashGenerator class to FabIL compiler for emitting class hashes in translated Fabric classes and interfaces.  The hash of a class/interface covers the text in its source file as well as the hashes of its super class (if any) and any implemented/extended interfaces.
XXX If the type of any class/interface is loaded from a signature or a .class file, then the Polyglot type information is used in lieu of the source file text.  We should be hashing the bytecode for the class/interface instead, but I don't know how to get that.

As it stands, our hashes appear to be stable across clean rebuilds of Polyglot, Jif, and Fabric.  (Stability across Java versions and across platforms is unknown.)