NamespaceChecker updates and simplifications. Eliminates cached field in FClass and is generally simpler to understand.  For namespace collisions, the original algorithm built the complete set of files that would need to be relinked should the user choose to do so.
I'm disabling this code for now, primarily to avoid having to update it.  We may
revisit this down the road once we understand better the kinds of feedback that
are useful.  For now, a namespace collision will result in a message like: 

"Detected namespace inconsistency for source Name.fab. The source's codebase
resolves pkg.DepName to a class object in 'codebase1 URL' but a dependency,
reached through pkg.DirectDepName, resolves pkg.DepName to a class object in
'codebase2 URL'.  If this is intentional, consider using an explicit codebase to
link with pkg.DirectDepName."

Auction example works.  Working on unit tests to excercise this explicitly.