ArgLabels are weird.  The declared label is actually in ArgLabel.upperBound(). I didn't spend enough time to understand exactly why ArgLabel needs to do this.
At any rate, containsThisLabel does not anticipate that a ThisLabel can be
in the upperBound() of an ArgLabel, so RemoteCallWrapperUpdater goes ahead and
generates a remote wrapper even though the signature contains a "this" label,
which is not representable at runtime (specifically, it is disallowed from
appearing in a label expression).  

Fixing containsThisLabel causes RemoteCallWrapperUpdater to (correctly) remove
remote wrappers for calls with {this} in the argument labels, as the code
implies it intended to do, even though it had no hope of doing so (bless its
heart).