Fabric requires worker$ to have the same LocalInstance everywhere it occurs. This ensures that Jif treats it as the same dynamic principal when it occurs in labels.
This is a problem when types are serialized. Occurrences of worker$ in type
signatures caused a different instance of worker$ to be used for checking labels
in signatures. 

Fortunately, Polyglot's type serialization is prepared for this. Since worker$
instances are expected to be == equal across separately compiled files, it
should be Internable.  When TypeInputStream sees a deserialized object that implements
Internable, it calls the intern() method.  This method returns an object that
replaces the deserialized one--in our case, ts.workerLocalInstance().