This commit improves coercion of workers and stores to principals. Additionally, it refactors some aspects of translation to java in an attempt to avoid a long-standing Polyglot codegen bug.  It is not always safe to expand class names to their fully-qualified form since the package name could conflict with in-scope variable names in some cases.
This update goes most of the way to avoiding expanding short names to FQ names,
but since the FabIL -> Java transformation is done in-place, by the time we get
to translation there exist many AmbPrefix nodes where type names have been
re-written to their _Impl names.  The Ambiguous nodes kill the typing context, so
we can't check the context to see whether it is safe to emit a short name.
Consequently, all names are expanded.  :(

It might be possible to override and hack the Translator to preserve the
context, but a more principled fix would be to refactor the FabIL -> Java
translation to use an ExtensionRewriter.