Small, warm T1 traversals on localhost now at 5.3s, down from 7.0s.
A slightly more intelligent object-grouping algorithm. When creating a group, we now traverse a bit further to ensure the object sub-graphs on the frontier of the group are large enough to form groups of their own. If not, we add those sub-graphs to the group being created.

Creating small OO7 DBs by default again.

TODO: make sure we aren't cheating -- check object group sizes to make sure they're reasonable.