Slightly less logging.  Small step towards a fix of read locking in semantic warranty transactions.
This message is mostly for me to note in the svn log what I need to fix next...

There's currently an issue where we are passing calls used back up to parent
logs, even in the case where we're making a new request.  This is *almost
certainly* related to the reason that I'm running into deadlocks with my
invalidating test in memo-top-item and probably would cause incorrect behavior
in defense (I don't think it would cause bugs, just overzealous additions of
dependencies when subcall warranties expire).