Class-level access policy syntax + better access policy checking.
This commit makes explicit the requirement that access policies may not contain
dynamic elements like fields or method arguments.  Some of these elements were
sneaking in via the default access policy assignment which tries to infer a
conservative access policy from the begin labels of methods or field labels.

In addition, additional syntax is provided to define the access policy for a
class or interface independently from fields.  This is especially important for
interface definitions.

Noticeably absent from this commit are several important features:
  code to enforce consistency between method begin labels and access policies.
  code to enforce consistency between interfaces and classes.
  other checks we have not yet considered.

Some of these problems were already latent in the code, but hopefully this
commit will make it easier to address these.

This commit was tested against friendmap and auction.