Truthified allocation label checks and a more complete implementation of store$ instantiation.
We are now checking that object and access labels flow to stores for
allocations with implicit locations.  Prior to this commit, we only checked
'new C@store()' expressions.

This commit should also fix most of the outstanding issues with the store$
pseudo-label.  In particular, instantiation of 'this.store$' expressions in
method signatures, should work better. 

This works in constructor signatures too by means of a new AccessPath for New
AST nodes. This access path is the mechanism that transfers the location
expression from the allocation site to the 'store$' occurrences.

Also fixed a subtle bug in label checking new expressions.  The reference label
of the allocated object was not being instantiated.