RFC: A mechanism for hashing references without fetching.
This provides support at the language level for hashing fabric.lang.Object
references.  Aside from being useful for performance, this ends up being
necessary for implementing delegating principals natively in Fabric. 

Consider adding a new principal to the set that defines the delegations of the
delegating principal.  Before the principal is added to the set, it doesn't act
for the delegating principal and thus the pc may not flow to the princpal's
store. Therefore, we want to implement the set using a comparator that can add items
without fetching them.