Make sif.servlet.Action a Fabric class.
This commit refactors the servlet code to make Action extend fabric.lang.Object
and moves the implementation into Fabric.  This paves the way for mobile Action
classes.

Examples auth, sif-hello, and groups are refactored to work with the new
changes.  The travel example compiles but something is screwy with the trust
relationships that I don't understand yet.