Refactored a bit. All read-map-related code should now be encapsulated in ReadMap.
Additionally, when we receive an object update, any existing read-map entry for the object now becomes defunct, and a new entry is created for the incoming update. This should prevent a race between updating of a read-map entry for a just-committed write and updating of the same entry for the resulting incoming update.

(Untested; committing so the Eucalyptus scripts will pick up these changes.)