Changed WarrantiedCallResult to carry around the set of oids for creates in the transaction that produced that result, so that later we can make a method for creating a fresh copy of the result for call reuse.
Changed the ReuseCallMessage to avoid duplicating transmitting the creates set.

Changed the creates collection on the store to collect creates of subcalls
(needed for full duplication of the result).

Changed the copy constructory methods to use LongSet and LongKeyMap to avoid
hash collisions.

Learned, once again, that I should commit more often in svn.  Curious that I
have the right habit for git but not for svn...