bugfix (please review). Don't use emptyList() for a mutable list
This commit fixes an issue with the SimpleSurrogateManager (SSM) trying to add
surrogates to a PrepareRequest's creates list.  If that list was size 0,
PrepareTransactionMessage used Collections.emptyList().  This resulted in an
UnsupportedOperationException when SSM tried to add items. See line 110 in SSM.

This seemed like safe fix, but perhaps there is a more efficient way.  For
instance, is it safe for SSM to allocate a new list and update req.creates?