Fixed issues where call checking for write prepares could have been using inconsistent data.
Still have an issue where transactions could be writing data used by calls that
they also make use of elsewhere in the call (not sure what the right thing to
do here is).

Started to amend issues where onum was confused with oid.  For now, it doesn't
affect any of our tests, so I'm putting that on a sticky note to correct later
when I have all the time in the world.