Bug fix: now avoiding infinite recursion when figuring out what the principals to whom an acts-for query may be forwarded.
Performance fix: when answering whether p≽q, only use delegations of the
form p≽a or a≽q. Previously, we introduced obligations p≽a and b≽q to
attempt to use a≽b delegations, causing duplication of work.

Tweaked some signatures to help ensure that we only forward acts-for
queries to PrimitivePrincipals.

Added PrincipalUtil.flowsTo(), a convenience method for deciding
information-flow relationships between principals.

Implemented toString() for ProofSearchState and ActsForQuery for easier
debugging.