A bit of a "hack" fix for a relatively rare concurrency bug.
Issue was that sometimes a writer would acquire a write lock on an Impl
that was evicted from cache but the transaction was still holding a
reference to.  Later when the transaction goes to write again through a
new reference (now pointing to the cached impl) this would result in an
attempt to take ownership from the local worker.  Since this only occurs
when the transction had been using an evicted impl, restart the entire
transaction.