Fixed locking protocol and support for modifying newly created treaties.
Locking now optimistically runs the first time and locks, not created in
the attempt, are all acquired before a retry.  This better allows for
identifying and grabbing all locks in the treaty tree that are likely
needed on the first attempt.

Also, put back in locking of full subtree (not just spine that is being
modified) to better avoid starvation on popular items.