Various changes to improve estimates, strategy, and extensions.
Clean up coming in a later commit, checking in as this is a large number
of changes already.

Estimates are back to a simple EWMA for time, velocity, and noise.

Strategy now actually uses an optimization library to attempt to select
starting slack to set expectation of 1 / time to violate treaty equal
across nodes.

Async extensions now run with a larger window of time than in-lined
extensions, to attempt to avoid coordination from updating treaties on
reads.