Strip down the implementation of BDB to address some bad interactions between the Fabric Store locks and the bdb locks.
This strips out a lot of pending transaction tables, which probably were
meant to help with recovering partial transactions after a failure.
However, we never implemented recovery protocols to use it and removing
it doesn't make the failure tolerance/recovery story any better/worse
than the MemoryDB implementation.  So, for now, let's run with the
simplified implementation.

This should address #36, will close if I confirm it works after some
testing.